Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpclient/trunk@1055732 13f79535-47bb-0310-9956-ffa450edef68

+++ httpclient-cache/src/main/java/org/apache/http/impl/client/cache/AsynchronousValidator.java
@@ -71,7 +71,7 @@ class AsynchronousValidator {
         this(cachingClient,
                 new ThreadPoolExecutor(config.getAsynchronousWorkersCore(),
                         config.getAsynchronousWorkersMax(),
-                        (long)config.getAsynchronousWorkerIdleLifetimeSecs(),
+                        config.getAsynchronousWorkerIdleLifetimeSecs(),
                         TimeUnit.SECONDS,
                         new ArrayBlockingQueue<Runnable>(config.getRevalidationQueueSize()))
                 );
